refactor(utils): split monolithic utils into focused submodules
Reorganize the sprawling utils.py into purpose-specific modules:
- file_ops, text_processing, legal_reasoning, formatting, core, logging_utils
- Update all command modules and tests to import from new locations
- No functional changes, purely structural cleanup
