+++ litassist/commands/extractfacts.py
@@ -11,17 +11,23 @@ import os
 
 from litassist.config import get_config
 from litassist.prompts import PROMPTS
-from litassist.utils import (
-    chunk_text,
-    save_log,
+from litassist.utils.text_processing import chunk_text
+from litassist.utils.file_ops import validate_file_size
+from litassist.utils.core import (
     timed,
-    create_reasoning_prompt,
-    save_command_output,
     show_command_completion,
+)
+from litassist.utils.legal_reasoning import (
+    create_reasoning_prompt,
+    verify_content_if_needed,
+)
+from litassist.utils.formatting import (
     info_message,
     success_message,
-    validate_file_size,
-    verify_content_if_needed,
+)
+from litassist.logging_utils import (
+    save_log,
+    save_command_output,
 )
 from litassist.llm import LLMClientFactory
 from litassist.verification_chain import run_cove_verification
