+++ tests/unit/test_command_parameters.py
@@ -86,7 +86,7 @@ class TestCommandParameterPropagation:
         )
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     def test_digest_summary_command_parameters(self, mock_read, mock_factory):
         """Test digest-summary command uses correct parameters."""
         mock_factory.return_value = self.mock_client
@@ -118,7 +118,7 @@ class TestCommandParameterPropagation:
         assert self.mock_client.complete.called
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     def test_digest_issues_command_parameters(self, mock_read, mock_factory):
         """Test digest-issues command uses correct parameters."""
         mock_factory.return_value = self.mock_client
@@ -199,7 +199,7 @@ class TestCommandParameterPropagation:
         assert LLMClientFactory.COMMAND_CONFIGS["lookup"]["force_verify"] is False
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     @patch("litassist.citation_verify.verify_all_citations")
     @patch("litassist.citation_patterns.extract_citations")
     def test_verify_command_parameters(
@@ -227,7 +227,7 @@ class TestCommandParameterPropagation:
         assert mock_factory.called
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     def test_brainstorm_command_parameters(self, mock_read, mock_factory):
         """Test brainstorm command uses correct parameters."""
         # Create a verification client for unorthodox verification
@@ -293,7 +293,7 @@ class TestCommandParameterPropagation:
         assert calls[3][0][0] == "brainstorm"  # analysis
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     def test_strategy_command_parameters(self, mock_read, mock_factory):
         """Test strategy command uses o3-pro model."""
         mock_factory.return_value = self.mock_client
@@ -348,7 +348,7 @@ Test objectives""")
                 mock_prompts.get_prompt.return_value = "Test prompt"
 
                 # Mock parse_strategies_file to avoid parsing issues
-                with patch("litassist.utils.parse_strategies_file") as mock_parse:
+                with patch("litassist.utils.core.parse_strategies_file") as mock_parse:
                     mock_parse.return_value = []
 
                     # Mock case facts validation
@@ -401,7 +401,7 @@ Test objectives""")
         assert LLMClientFactory.COMMAND_CONFIGS["strategy"]["force_verify"] is True
 
     @patch("litassist.llm.LLMClientFactory.for_command")
-    @patch("litassist.utils.read_document")
+    @patch("litassist.utils.file_ops.read_document")
     @patch("litassist.helpers.retriever.get_pinecone_client")
     @patch("litassist.commands.draft.get_config")
     def test_draft_command_parameters(
